population-service/src/requirements.txt
grpcio>=1.28.1
protobuf>=3.5.0.post1
pika>=1.1.0
qwikidata>=0.4.0
fluent-logger>=0.9.6